Add user name and picture to html templates
When the user logs into the page, the nav bar will include their name
and picture with a log out button. When the user is not logged in there
is a button that takes you to the login page. The root path in
itemCatalog.py was changed to be the same path as '/breweries'