Fix two bugs
First Bug: Logged in User info was not displayed on create_brewery.html
and updateBrewery.html. An integer was being supplied as "user" instead
of the object "user"

Second Bug: User that created the initial entries in the database was
not giving allowing that user access to updateBrewery.html to make a
change. An integer for current user id was being evaluated against the
entire object of the creator not the creator id.